requirements/requirements.txt
asyncio
deepspeed>=0.7.6
deepspeed-kernels
Flask-RESTful
grpcio
grpcio-tools
pydantic
torch
transformers
ujson
Werkzeug
zmq
